+++ packages/admin/cms-admin/src/documents/types.ts
@@ -1,4 +1,5 @@
 import { TypedDocumentNode } from "@apollo/client";
+import { BlockDependency, ReplaceDependencyObject } from "@comet/blocks-admin";
 import { SvgIconProps } from "@mui/material";
 
 import { GQLDocumentInterface, Maybe } from "../graphql.generated";
@@ -31,8 +32,6 @@ export interface GQLDocument extends GQLDocumentInterface {
     [key: string]: unknown;
 }
 
-export type IdsMap = Map<string, string>;
-
 export interface DocumentInterface<
     DocumentInput extends Record<string, unknown> = Record<string, unknown>,
     DocumentOutput extends Record<string, unknown> = Record<string, unknown>,
@@ -41,9 +40,11 @@ export interface DocumentInterface<
     getQuery?: TypedDocumentNode<GQLPageQuery, GQLPageQueryVariables>; // TODO better typing (see createUsePage.tsx)
     editComponent?: React.ComponentType<{ id: string; category: string }>;
     updateMutation?: TypedDocumentNode<GQLUpdatePageMutation, GQLUpdatePageMutationVariables<DocumentInput>>;
-    inputToOutput?: (input: DocumentInput, context: { idsMap: IdsMap }) => DocumentOutput;
+    inputToOutput?: (input: DocumentInput) => DocumentOutput;
     menuIcon: (props: SvgIconProps<"svg">) => JSX.Element | null;
     hideInMenuIcon?: (props: SvgIconProps<"svg">) => JSX.Element | null;
     InfoTag?: React.ComponentType<{ page: PageTreePage }>;
     anchors: (input: DocumentInput) => string[];
+    dependencies: (input: DocumentInput) => BlockDependency[];
+    replaceDependenciesInOutput?: (output: DocumentOutput, replacements: ReplaceDependencyObject[]) => DocumentOutput;
 }